# generated/server/upgrade/upgrade_main.py
import uvicorn; from fastapi import FastAPI, HTTPException; from fastapi.middleware.cors import CORSMiddleware; from pydantic import BaseModel
app = FastAPI(title='Enterprise Core Upgrade Engine', version='2.0.0')
app.add_middleware(CORSMiddleware, allow_origins=["*"], allow_methods=["*"], allow_headers=["*"])
current_cluster_versions = {"gateway": "2.0.0", "security": "2.0.0", "migration": "2.0.0", "config": "2.0.0", "upgrade": "2.0.0", "fencing": "2.0.0", "analytics": "2.0.0"}
class UpgradePayload(BaseModel):
    service_name: str
    target_version: str
    force_update: bool = False
@app.get('/health')
def health(): return {'status': 'healthy', 'node': 'upgrade', 'port': 8060, 'registry': current_cluster_versions}
@app.post('/api/v1/upgrade/apply')
def apply_upgrade(payload: UpgradePayload):
    if payload.service_name not in current_cluster_versions: raise HTTPException(status_code=404, detail="Service node not found")
    old_v = current_cluster_versions[payload.service_name]
    current_cluster_versions[payload.service_name] = payload.target_version
    return {"status": "upgraded", "service": payload.service_name, "previous_version": old_v, "current_version": payload.target_version, "strategy": "Zero-Downtime Rolling Patch"}
if __name__ == '__main__': uvicorn.run(app, host='0.0.0.0', port=8060)